Expand args to absolute paths before checking their validity.
This fixes a bug in which the validity of FROM_DIR and TO_DIR were
checked before ~ and relative paths in those arguments were expanded,
sometimes this could cause dotfilemanager to complain that a valid path
was invalid. Expand the paths and _then_ check their validity.

You can now just do `dotfilemanager tidy` (with no further arguments) to
tidy up symlinks in your homedir, and `dotfilemanager report` or
`dotfilemanager link` to report on or make links from ~ into
~/.dotfiles.